// frontend/src/redux/features/types.d.ts
export type TypeNewConversation={
    _id: string;
    phone_number: string;
    status: string;
    chat_ids: string[];
    credential_id: string;
    createdAt: string;
    updatedAt: string;
    __v: 0;
  }